Chore [#19] Mark 주석 추가
