+++ YELLO-iOS/YELLO-iOS/Presentation/Onboarding/FriendModel.swift
@@ -1,0 +1,14 @@
+//
+//  FriendModel.swift
+//  YELLO-iOS
+//
+//  Created by 지희의 MAC on 2023/07/13.
+//
+
+import Foundation
+
+struct FriendModel {
+    let name: String
+    let school: String
+    var isButtonSelected: Bool
+}
