+++ YELLO-iOS/YELLO-iOS/Presentation/Onboarding/ViewController/Base/SearchBaseViewController.swift
@@ -7,18 +7,24 @@
 
 import UIKit
 
+// MARK: - Delegate
 protocol SearchResultTableViewSelectDelegate: AnyObject {
     func didSelectSearchResult(_ result: String)
 }
 
 class SearchBaseViewController: BaseViewController {
     
+    // MARK: - Variables
+    // MARK: Property
     var allArr: [String] = []
     var filterdArr: [String] = []
     
+    // MARK: Component
     let searchView = SearchView()
     weak var delegate: SearchResultTableViewSelectDelegate?
     
+    // MARK: - Function
+    // MARK: - LifeCycles
     override func loadView() {
         super.loadView()
         view = searchView
@@ -35,6 +41,7 @@ class SearchBaseViewController: BaseViewController {
         setDelegate()
     }
     
+    // MARK: Custom Function
     func customView(titleText: String, helperText: String, allArr: [String]) {
         searchView.titleLabel.text = titleText
         searchView.helperButton.setTitle(helperText, for: .normal)
@@ -53,10 +60,12 @@ class SearchBaseViewController: BaseViewController {
         searchView.searchResultTableView.delegate = self
     }
     
+    /// 키보드 내리기
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         self.view.endEditing(true)
     }
     
+    // MARK: Objc Function
     @objc func textFieldDidChange(_ textField: UITextField) {
         filterdArr = allArr.filter { search in
             return search.contains(textField.text ?? "")
@@ -70,6 +79,8 @@ class SearchBaseViewController: BaseViewController {
     
 }
 
+// MARK: - extension
+// MARK: UITextFieldDelegate
 extension SearchBaseViewController: UITextFieldDelegate {
     
     func updateSearchResults(for searchController: UISearchController) {
@@ -80,6 +91,8 @@ extension SearchBaseViewController: UITextFieldDelegate {
     
 }
 
+// MARK: - extension
+// MARK: UITableViewDataSource
 extension SearchBaseViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         filterdArr.count
@@ -87,13 +100,16 @@ extension SearchBaseViewController: UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let searchResult = filterdArr[indexPath.row]
-        let cell = tableView.dequeueReusableCell(withIdentifier: "SearchResultTableViewCell") as! SearchResultTableViewCell
+        let cell = tableView.dequeueReusableCell(withIdentifier: SearchResultTableViewCell.identifier) as! SearchResultTableViewCell
         cell.titleLabel.text = searchResult
         cell.selectionStyle = .none
         return cell
     }
 }
 
+// MARK: - extension
+
+// MARK: UITableViewDelegate
 extension SearchBaseViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         guard let currentCell = tableView.cellForRow(at: indexPath) as? SearchResultTableViewCell else {
