+++ YELLO-iOS/YELLO-iOS/Presentation/Onboarding/Views/Helper/YelloButton.swift
@@ -6,28 +6,34 @@
 //
 
 import UIKit
-import SnapKit
 
+import SnapKit
+import Then
+
+// MARK: - enum
+/// 버튼 상태
 enum ButtonState {
     case enabled
     case unabled
 }
 
 final class YelloButton: UIButton {
-    
+    // MARK: - Variables
+    // MARK: Constants
     let roundAmount: CGFloat = 8
     let buttonHeight: CGFloat = 48
     let textSize: CGFloat = 15
+    
+    // MARK: Property
     var buttonState: Bool = false
     var buttonText: String?
     
+    // MARK: - Function
+    // MARK: LifeCycle
     override init(frame: CGRect) {
         super.init(frame: frame)
     }
     
-    // MARK: - Function
-
-    // MARK: LifeCycle
     init(buttonText: String, isEnabled: Bool = false) {
         super.init(frame: CGRect(x: 0, y: 0, width: buttonHeight, height: 0))
         self.buttonText = buttonText
@@ -41,19 +47,28 @@ final class YelloButton: UIButton {
     }
 }
 
+// MARK: - extension
+// MARK: setUI
 extension YelloButton {
-    private func setUI() {
+    func setUI() {
         setStyle()
         setLayout()
     }
     
     private func setStyle() {
         self.setTitle(buttonText, for: .normal)
-        self.makeCornerRound(radius: roundAmount)
+        self.makeCornerRound(radius: CGFloat(Constraints.round))
         self.titleLabel?.font = .uiButton
         setButtonEnable(state: self.buttonState)
     }
     
+    private func setLayout() {
+        self.snp.makeConstraints {
+            $0.height.equalTo(buttonHeight)
+        }
+    }
+    
+    // MARK: Custom Method
     func setButtonEnable(state: Bool) {
         self.isEnabled = state
         switch self.isEnabled {
@@ -63,12 +78,6 @@ extension YelloButton {
         case false:
             self.backgroundColor = .grayscales800
             self.setTitleColor(.grayscales700, for: .normal)
-        }
-    }
-    
-    private func setLayout() {
-        self.snp.makeConstraints {
-            $0.height.equalTo(buttonHeight)
         }
     }
 }
